Update README.md
# Leaf Segmentation

# Project Overview
We have to segment leaves from image of plant. For that we have to first detect the leaves and after detection of leaves find leaf area from image of plant.
After finding leaf area we have to find edges of leaves. Now we have to segment a leaves on the basis of edges of leaves. 

# Problem and Solution Statement
Segmentation of leaf instances in the UAV(Unnamed Aerial Vehicles) images and it can be used to measure  various phenotypic traits such as leaf length, maximum leaf width, and leaf area in-dex. Successful leaf segmentation accurately detects leaf edges.

# Benchmark (How this solution is better?)
Our main motivation behind this project is as this definition is open research challenge by CodaLab and after  segmenting the leaf we are able to find numbers of leaves in plant and using that we are able to find leaf length, width and area using this factors of leaves we are able to predict the growth of plants which is most necessary thing in agriculture.

# Implementation strategy
Image based plant phenotyping is a growing application field of computer vision in agriculture. A major function is the segmentation of all individual leaves into images. Here we focus on the most common rosette model plants, Arabidopsis and young tobacco. Although leaves exhibit shared form and shape characteristics, the presence of leaf shape and posture improvement and variability, as well as imaging conditions, challenge this problem.